feat: 메뉴 패키지를 모듈로 별도 분리하라
