+++ module-menu/src/main/java/kitchenpos/menu/application/validator/MenuValidator.java
@@ -1,0 +1,7 @@
+package kitchenpos.menu.application.validator;
+
+import kitchenpos.menu.dto.request.MenuRequest;
+
+public interface MenuValidator {
+    void execute(MenuRequest menuRequest);
+}
