+++ module-menu/src/main/java/kitchenpos/menu/dto/request/MenuProductRequest.java
@@ -1,0 +1,26 @@
+package kitchenpos.menu.dto.request;
+
+import kitchenpos.menu.domain.MenuProduct;
+
+public class MenuProductRequest {
+
+    private Long productId;
+    private long quantity;
+
+    public MenuProductRequest(Long productId, long quantity) {
+        this.productId = productId;
+        this.quantity = quantity;
+    }
+
+    public MenuProduct toEntity() {
+        return MenuProduct.of(productId, quantity);
+    }
+
+    public Long getProductId() {
+        return productId;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+}
