+++ module-menu/src/main/java/kitchenpos/menu/dto/request/MenuRequest.java
@@ -1,0 +1,46 @@
+package kitchenpos.menu.dto.request;
+
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.stream.Collectors;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
+
+public class MenuRequest {
+
+    private String name;
+    private BigDecimal price;
+    private Long menuGroupId;
+    private List<MenuProductRequest> menuProductRequests;
+
+    public MenuRequest(String name, BigDecimal price, Long menuGroupId, List<MenuProductRequest> menuProductRequests) {
+        this.name = name;
+        this.price = price;
+        this.menuGroupId = menuGroupId;
+        this.menuProductRequests = menuProductRequests;
+    }
+
+    public Menu toEntity() {
+        List<MenuProduct> menuProducts = this.menuProductRequests.stream()
+                .map(MenuProductRequest::toEntity)
+                .collect(Collectors.toList());
+
+        return Menu.of(this.name, this.price, this.menuGroupId, menuProducts);
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public Long getMenuGroupId() {
+        return menuGroupId;
+    }
+
+    public List<MenuProductRequest> getMenuProductRequests() {
+        return menuProductRequests;
+    }
+}
