+++ module-menu/src/main/java/kitchenpos/menu/dto/response/MenuResponse.java
@@ -1,0 +1,60 @@
+package kitchenpos.menu.dto.response;
+
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.stream.Collectors;
+import kitchenpos.menu.domain.Menu;
+
+public class MenuResponse {
+
+    private Long id;
+    private String name;
+    private BigDecimal price;
+    private Long menuGroupId;
+    private List<MenuProductResponse> menuProducts;
+
+    public MenuResponse() {
+    }
+
+    private MenuResponse(Long id, String name, BigDecimal price, Long menuGroupId,
+        List<MenuProductResponse> menuProducts) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+        this.menuGroupId = menuGroupId;
+        this.menuProducts = menuProducts;
+    }
+
+    public static MenuResponse of(Menu menu) {
+        List<MenuProductResponse> menuProducts = menu.getMenuProducts().stream()
+            .map(MenuProductResponse::of)
+            .collect(Collectors.toList());
+
+        return new MenuResponse(
+            menu.getId(),
+            menu.getName(),
+            menu.getPrice(),
+            menu.getMenuGroupId(),
+            menuProducts);
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public Long getMenuGroupId() {
+        return menuGroupId;
+    }
+
+    public List<MenuProductResponse> getMenuProducts() {
+        return menuProducts;
+    }
+}
