+++ module-menu/src/test/java/kitchenpos/menu/ModuleMenuApplicationTests.java
@@ -1,0 +1,11 @@
+package kitchenpos.menu;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ModuleMenuApplicationTests {
+    @Test
+    public void contextLoads() {
+    }
+}
