+++ module-menu/src/test/resources/application.properties
