+++ module-menugroup/src/main/java/kitchenpos/menugroup/domain/MenuGroupRepository.java
@@ -1,6 +1,8 @@
 package kitchenpos.menugroup.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
 }
