+++ settings.gradle
@@ -5,4 +5,5 @@ include 'module-order'
 include 'module-core'
 include 'module-product'
 include 'module-menugroup'
+include 'module-menu'
