--- src/main/java/kitchenpos/menu/application/validator/MenuValidator.java
@@ -1,7 +1,0 @@
-package kitchenpos.menu.application.validator;
-
-import kitchenpos.menu.domain.request.MenuRequest;
-
-public interface MenuValidator {
-    void execute(MenuRequest menuRequest);
-}
