--- src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,79 +1,0 @@
-package kitchenpos.menu.domain;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.stream.Collectors;
-import javax.persistence.Embedded;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.Table;
-import kitchenpos.common.domain.Name;
-import kitchenpos.common.domain.Price;
-import kitchenpos.menu.domain.request.MenuProductRequest;
-import kitchenpos.menu.domain.request.MenuRequest;
-
-@Entity
-@Table(name = "menu")
-public class Menu {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Embedded
-    private Name name;
-
-    @Embedded
-    private Price price;
-
-    @JoinColumn(name = "menu_group_id")
-    private Long menuGroupId;
-
-    @Embedded
-    private MenuProducts menuProducts = new MenuProducts();
-
-    protected Menu() {
-    }
-
-    private Menu(String name, BigDecimal price, Long menuGroupId, List<MenuProduct> menuProducts) {
-        this.name = new Name(name);
-        this.price = new Price(price);
-        this.menuGroupId = menuGroupId;
-        this.menuProducts = new MenuProducts(menuProducts, this);
-    }
-
-    public static Menu of(String name, BigDecimal price, Long menuGroupId, List<MenuProduct> menuProducts) {
-        return new Menu(name, price, menuGroupId, menuProducts);
-    }
-
-    public static Menu of(MenuRequest menuRequest) {
-        List<MenuProduct> menuProducts = menuRequest.getMenuProductRequests().stream()
-            .map(MenuProductRequest::toEntity)
-            .collect(Collectors.toList());
-
-        return new Menu(menuRequest.getName(), menuRequest.getPrice(), menuRequest.getMenuGroupId(), menuProducts);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name.getValue();
-    }
-
-    public BigDecimal getPrice() {
-        return price.getValue();
-    }
-
-    public Long getMenuGroupId() {
-        return menuGroupId;
-    }
-
-    public List<MenuProduct> getMenuProducts() {
-        return menuProducts.getItems();
-    }
-}
