--- src/main/java/kitchenpos/menu/domain/response/MenuProductResponse.java
@@ -1,38 +1,0 @@
-package kitchenpos.menu.domain.response;
-
-import kitchenpos.menu.domain.MenuProduct;
-
-public class MenuProductResponse {
-    private Long seq;
-    private Long productId;
-    private long quantity;
-
-    public MenuProductResponse() {
-    }
-
-    public MenuProductResponse(Long seq, Long productId, long quantity) {
-        this.seq = seq;
-        this.productId = productId;
-        this.quantity = quantity;
-    }
-
-    public static MenuProductResponse of(MenuProduct menuProduct) {
-        return new MenuProductResponse(
-            menuProduct.getSeq(),
-            menuProduct.getProductId(),
-            menuProduct.getQuantity()
-        );
-    }
-
-    public Long getSeq() {
-        return seq;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-}
