--- src/test/java/kitchenpos/AcceptanceTest.java
@@ -1,25 +1,0 @@
-package kitchenpos;
-
-import io.restassured.RestAssured;
-import kitchenpos.utils.DatabaseCleanUp;
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.web.server.LocalServerPort;
-
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-public class AcceptanceTest {
-
-    @LocalServerPort
-    int port;
-
-    @Autowired
-    private DatabaseCleanUp databaseCleanUp;
-
-    @BeforeEach
-    public void setUp() {
-        RestAssured.port = port;
-        databaseCleanUp.execute();
-    }
-}
